+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### UNRELEASED
-* fix outbound ip detection
+* bootstrap: fix outbound ip detection
+* bootstrap: remove unnecessary double close error log 
 ### v2.8.1
 * fix release tag
 ### v2.8.0